chore: Updated dependencies
Notable changes:
Mincer, which is exported, upgraded to ^2.0.1
For some reason there's a test for the status code http.get returns when there is '..' in the url. This now returns a 404 instead of a 400.
argparser upgraded to ^2.0.1. Only used in bin/connect-assets and seems mostly compatible with previous versions, with a few renames needed.
Other upgrades are for dev dependencies or just makes it match the versioning in other modules.

Audits still failing:
lodash version included by mincer
Various mocha dependencies, but mocha is a devDependency